expose Unmap methods on Buffer and Memory for usage after calling Map ; update examples; optimize MapInfo.WriteData to use memcpy
